+++ heksher/clients/thread_client.py
@@ -78,9 +78,9 @@ class ThreadHeksherClient(V1APIClient, ContextFeaturesMixin, ContextManagerMixin
         http_client = self._http_client()
 
         def declare_setting(setting):
-            response = http_client.put('api/v1/settings/declare',
-                                       content=orjson.dumps(setting.to_v1_declaration_body()),
-                                       headers=content_header)
+            response = http_client.post('api/v1/settings/declare',
+                                        content=orjson.dumps(setting.to_v1_declaration_body()),
+                                        headers=content_header)
             self._handle_declaration_response(setting, response)
 
         while self._keep_going: